Allow passing integers as tags to TaggedLogging::Formatter
Logstash and Elasticsearch accept integer values for keyword type
fields, such as tags. LogStashLogger filters tags, not to pass nils
or empty strings as tags to logstash. However, the empty string check
does fail on everything that is not a string, and it does so in not a
very friendly manner.
This PR makes that check happen only on objects that respond to #empty?.